app/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis para cola de mensajes
redis==5.0.1

# Alpaca trading
alpaca-trade-api==3.1.1

# Validación y configuración
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilidades
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSockets (compatible con alpaca-trade-api)
websockets>=9.0,<11

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

pydantic[email]==2.2.0